// client/src/constants/cards-types.ts
export enum CardsType {
  Maple = "maple",
  Sakura = "sakura",
  Pine = "pine",
  Iris = "iris",
  Sun = "sun",
  Rain = "rain",
  Bird = "bird",
  Paper = "paper",
}

export const RU_CARDS_TYPES = {
  [CardsType.Maple]: "клён",
  [CardsType.Sakura]: "сакура",
  [CardsType.Pine]: "сосна",
  [CardsType.Iris]: "ирис",
  [CardsType.Sun]: "солнце",
  [CardsType.Rain]: "дождь",
  [CardsType.Bird]: "птица",
  [CardsType.Paper]: "тандзаку",
};